Repaired repo
This commit incorporates a series of changes made up to and including
the accidental addition of large data files that necessitated cloning
the existing repo again. Here are major changes:

Added a "run" method to SimpleRC that feeds output back into the input
of an RC to allow it to predict a timeseries given an initial condition.

Added scripts to use the run method for kuramoto and weather.

Changed the .gitignore file.

Fixed an error in the normalization of weather data.

 Changes to be committed:
	modified:   .gitignore
	modified:   W_VA
	modified:   Wout_VA
	modified:   az_preds
	modified:   az_y_test
	new file:   kuramoto2.py
	modified:   kuramoto_output
	modified:   simpleRC.py
	modified:   va_preds
	modified:   va_y_test
	modified:   weather.py
	new file:   weather2.py
	modified:   weather_output