Core algorithm
Corrected an error in the "run" method that resulted in zeroing the
reservoir incorrectly. In doing so, implemented two versions of "run"
and "train" that allow the RC to be used in a "one-shot" mode (i.e., as
a simple transfer function that, from a zeroed out reservoir, gives the
right function of the input and as a forced input system that allows the
reservoir to evolve with the input such that the reservoir state after
an input depends on prior input.